requirements.txt
langgraph
langchain_openai
langchain_google_genai
langchain_community
tavily-python
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
gTTS
moviepy
huggingface_hub
diffusers
accelerate
safetensors
transformers
torch
